fix book delete permissions

+++ app/Http/Controllers/BookController.php
@@ -94,7 +94,7 @@ class BookController extends Controller
      */
     public function destroy(Book $book)
     {
-        // $this->authorize('delete', $book);
+        $this->authorize('delete', $book);
 
         $book->delete();
